# Modify Rack::Session::Dalli to be compatible with ConnectionPool