I tried do this with as small of a footprint as possible. Let me know if you'd like me to take a different approach, e.g. with argument checking, or eliminate the `with_lock` method entirely if it's a ConnectionPool. Thank you!

n.b. This addresses #562!