@mwpastore 2016-02-13
I don't know why that one test failed, but it appears to be a transient error:

```
failed to listen on TCP port 34547: Address already in use
failed to listen on TCP port 34541: Address already in use
```