@petergoldstein 2016-02-13
So I think is a reasonable first pass, but I think the initialization phase doesn't cover the cases we probably want it to cover and isn't entirely consistent with how connection pool support is implemented elsewhere.  Specifically here - https://github.com/petergoldstein/dalli/commit/c51ce5d826b3556cb9533ff3246277613d10dc2c

My take is that you probably want to handle four scenarios, in the following order:

1. Caller passes in pool_options which, if present, cause a ConnectionPool to be instantiated.
2. Caller passes in a ConnectionPool instance
3. Caller passes in a Dalli::Client instance
4. Caller passes in nothing, and the initialize method instantiates a Dalli::Client for internal use

The current form of the PR doesn't allow option 1, which seems incongruous considering that Dalli explicitly supports that in the Rails cache implementation.

I'd suggest breaking the `initialize` method into pieces to handle each of the 4 options above, and set up the @pool and @mutex variables as needed.

What do you think?