@mwpastore 2016-02-13
Let me start by saying that I think these are all good ideas and I'm certainly willing to put in the effort. 

However, it would involve quite a significant rewrite of Rack::Session::Dalli's constructors. Currently, the only attributes of the Dalli::Client object that can be customized when instantiating Rack::Session::Dalli are `:memcache_server` and `:namespace`. It doesn't support any of the other options, like `:failover` or `:serializer`; it looks like it was written to be explicitly, 1:1 compatible with Rack::Session::Memcache. Currently, the only way to customize the Dalli::Client object are to instantiate it yourself ahead of time and pass it in as the (undocumented? but widely referenced online) `:cache` option. To maintain a stable interface, the best option is to simply allow the user to pass in a ConnectionPool object instead of a Dalli::Client object, which is what I've done here. Rack::Session::Dalli has a completely different ancestry and use case than ActiveSupport::Cache::DalliStore so I'm not sure it makes sense to look there for a model implementation, but I certainly understand the desire to be consistent.

All that being said, if you're comfortable with me embarking on this overhaul, possibly changing the inheritance model, and possibly breaking compatibility with Rack::Session::Memcache, I will table this PR for now and come back to it in a few weeks when I have more time.