@mwpastore 2016-02-13
Argh, okay, let me walk that back a little. I didn't understand how the Abstract::ID parent class was manipulating `DEFAULT_OPTIONS` and `@default_options`. You definitely _can_ pass in e.g. `:compress` and it will get passed to Dalli::Client.new. Let me play with this a bit... maybe it won't be as bad as I think.