@petergoldstein 2016-02-13
Ok, I thought your initial comments were a little strong.  I'm unclear why you can't use default_options to handle the cases I describe, in a way that is completely backwards compatible, but supports the additional cases with other option flags.  That's essentially what's being done with Dalli::Client now.