@mwpastore 2016-02-13
They were a little strong, and I apologize. I'm a dope. Thanks Peter.