@mwpastore 2016-02-13
Here's a question that maybe you can answer: Why does this thing even have a mutex in the first place? Doesn't Dalli::Client do its own thread-safety? Is it possibly a holdover from before it was thread-safe (and therefore we can remove it)?