@petergoldstein 2016-02-13
That's actually an excellent question, and I don't really know the answer.  From `blame` we can see that it was added in PR #264 , which post-dates the introduction of Dalli thread safety.  From the comments in the PR, it appears to be driven by concerns over the removal of Rack::Lock in Rails 4.  And the code seems to be almost a straight port of https://github.com/rack/rack/blob/master/lib/rack/session/memcache.rb

But as you note, Dalli::Client is threadsafe by default, so my suspicion is that the mutex (and the corresponding complexity around it) is probably unnecessary.  And the same would be true with a ConnectionPool.

I don't have time to look into this right now, but I suspect there's a little cargo cultism going on here, and we could eliminate the locking logic entirely.