@mwpastore 2016-02-13
Something is still not quite right with this options parsing logic. I think there might have been a hint of truth to what I was trying to say in my mini-rant earlier. If you pass in e.g. `:compress => true` as an option to Rack::Session::Dalli, it will be present in `@default_options`, but [this line](https://github.com/petergoldstein/dalli/blob/v2.7.5/lib/rack/session/dalli.rb#L17) will exclude it from `mopts` (because it's not present in `DEFAULT_OPTIONS`), and it won't be passed to Dalli::Client.new. The filter should really be by the options that Dalli::Client.new accepts, plus the pool options we've just introduced. I'm working on it...