@mwpastore 2016-02-13
Okay, I'm feeling pretty confident about this. Take a look. I've added a documentation block as well, and I made some assertions in there that may not be true, so a sanity check would be appreciated. Thanks Peter!