@petergoldstein 2016-02-14
@mwpastore Appreciate the effort.  I'm unlikely to get a chance to review today, but I'll try to get to it this week.

Also, there have been some changes since this was pushed - Rails 5 compatibility issues.  Can you rebase this branch when you have a minute?

Thanks