@PetrKaleta 2016-02-22
This is exactly what I was looking for. I hope it will be merged soon