@petergoldstein 2016-03-06
@mwpastore Sorry it's taken me a few weeks to get back to this.

On a high level, I think the approach looks ok.  I left a few line comments that need to be addressed before we can discuss merging.

As per our earlier conversation, I suspect that the mutexes were there from a cargo cult change, and not because they were actually required.  So I'm fine with removing them, provided you do a bit of concurrency sanity checking beforehand.

It looks like you covered the cases I mentioned in my earlier comment, albeit in a slightly different order (2/3, 1, 4).  I think that ordering is probably fine.

One key item that's missing is any sort of testing.  You're going to need to add some tests to cover this new functionality, and to ensure options are being handled as expected.

Finally, I'd still recommend breaking up the initialize method or, at the very least, adding a couple of syntactic sugar methods to clarify exactly what is being tested and done in the `if` clause used to create the `@pool` object