@mwpastore 2016-03-06
@petergoldstein No worries. Thank you for taking the time today to review it. I will work on adding some additional tests for the new functionality and breaking up `initialize`.