@mwpastore 2016-03-07
@petergoldstein Alrighty, take another look when you can and let me know if you'd like to see any more changes or tests!