@mwpastore 2016-03-09
Hi Peter, just a heads up that I found and fixed a new bug yesterday. I had changed the calling signature of `generate_sid` to avoid nesting `@pool.with` blocks, not realizing that `generate_sid` is called directly from Rack::Session::Abstract::Persisted [here](https://github.com/rack/rack/blob/1.6.4/lib/rack/session/abstract/id.rb#L332) if `:renew` or `:drop` are specified. I think I found a solution that lets us have the best of both worlds; the relevant line changes are [here](https://github.com/mwpastore/dalli/blob/rack-session/lib/rack/session/dalli.rb#L149-L154). I also added a test to prevent this from tripping up someone else in the future. Let me know if you have any questions.