@mwpastore 2016-03-10
## Benchmark Results

Here are some preliminary benchmark results comparing the performance of Rack::Session::Dalli under heavy thread concurrency before and after this patch, and with and without connection pooling. Testing was performed using wrk against a simple Sinatra app running on Puma 3.1/JRuby 9.0.5.0/OpenJDK 7u95. The benchmark is far from comprehensive but it does a get and a set on every request. My code is [here](https://github.com/mwpastore/dalli-bench).

### Without Connection Pooling

#### Before (Dalli 2.7.6 release)

```
Running 1m test @ http://localhost:9292
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   354.10ms  101.44ms 976.56ms   69.19%
    Req/Sec    91.84     29.26   323.00     70.14%
  65961 requests in 1.00m, 15.23MB read
Requests/sec:   1097.82
Transfer/sec:    259.60KB
```

#### After

```
Running 1m test @ http://localhost:9292
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   381.95ms  137.87ms 969.73ms   69.78%
    Req/Sec    84.80     27.88   200.00     72.44%
  60879 requests in 1.00m, 14.06MB read
Requests/sec:   1012.99
Transfer/sec:    239.58KB
```

#### Delta

Avg Latency: 1.08x
Requests/sec: 0.92x

Unfortunately, it looks like this new version is about 8% slower than the current release (for this simple benchmark, at least). I'd like to write some different benchmarks and do some additional profiling but I think it's safe to assume that calling Dalli::Client#with with a block is just simply slower than calling the #get, #set, #add, and #delete methods directly. It probably would be even worse if we hadn't also removed the unnecessary mutex. 

### With Connection Pooling

```
Running 1m test @ http://localhost:9292
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    98.59ms   47.54ms 453.85ms   58.51%
    Req/Sec   336.16     48.45   707.00     74.34%
  241251 requests in 1.00m, 55.72MB read
Requests/sec:   4014.28
Transfer/sec:      0.93MB
```

#### Delta (cf. Dalli 2.7.6 release)

Avg Latency: 0.28x
Requests/sec: 3.96x

That's more like it! So the question becomes whether the the huge speedup with connection pooling justifies the slight slowdown without.