@mwpastore 2016-03-14
### with vs. without with

I wrote a simple benchmark to simulate a session flow for a single request with `:renew` set to true (which is what the main benchmark does) and a bad rack.session cookie on the request. Some of the ops will occur within the same call to with_block, others will take place in their own. Code is up on the same Git repository.

```
Warming up --------------------------------------
        without with   132.000  i/100ms
                with   152.000  i/100ms
Calculating -------------------------------------
        without with      1.672k (± 5.8%) i/s -     50.028k
                with      1.649k (± 6.4%) i/s -     49.248k
```

> I wouldn't assume this without testing

As you suspected, my assumptions did not bear out, so I will have to do some more profiling to figure out where the performance hit is coming from. #with only accounts for about 1.6% of the difference. 

### wrk revisited

> Are 400 connections even useful in a test with 12 threads?

I'm not sure. This is my first experience with wrk and the documentation is... scant. My understanding is that the threads mostly come into play when you are running Lua scripts to simulate more complex sequences of requests, and that simply juggling 400 open connections requires significantly less horsepower. wrk was never a huge consumer of CPU (compared to Ruby and Memcached) in my testing, even when there were ample cycles available for it to consume. 

I kept the threads at 12 and ran some tests to show the scaling. I couldn't do 10 because wrk requires at least as many connections as there are threads. I ran these on a differently-sized box than the above so I repeated the 400 connections test as well. Let me know what else you'd like to see. 

#### 12 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 12 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    40.21ms    1.30ms  60.02ms   93.26%
    Req/Sec    24.88      5.00    30.00     51.12%
  17914 requests in 1.00m, 4.14MB read
Requests/sec:    298.28
Transfer/sec:     70.53KB
```

#### 20 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    40.25ms    1.30ms  53.31ms   92.06%
    Req/Sec    24.85      5.00    30.00     51.43%
  17895 requests in 1.00m, 4.13MB read
Requests/sec:    297.96
Transfer/sec:     70.46KB
```

#### 40 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    40.72ms    1.79ms  60.07ms   79.61%
    Req/Sec    73.68      9.22    90.00     69.15%
  53053 requests in 1.00m, 12.25MB read
Requests/sec:    883.57
Transfer/sec:    208.95KB
```

#### 80 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 80 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    41.83ms    2.55ms  70.88ms   90.71%
    Req/Sec   143.90     13.12   181.00     62.00%
  103304 requests in 1.00m, 23.86MB read
Requests/sec:   1720.18
Transfer/sec:    406.79KB
```

#### 160 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 160 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    49.19ms    8.88ms 208.80ms   72.72%
    Req/Sec   264.05     27.33   350.00     77.64%
  189601 requests in 1.00m, 43.79MB read
Requests/sec:   3154.89
Transfer/sec:    746.10KB
```

#### 320 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 320 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    92.53ms   45.91ms 651.36ms   62.75%
    Req/Sec   282.63     44.76   470.00     72.11%
  202886 requests in 1.00m, 46.86MB read
Requests/sec:   3375.87
Transfer/sec:    798.40KB
```

#### 400 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 400 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   114.08ms   64.03ms 524.25ms   58.39%
    Req/Sec   291.81     53.37   717.00     73.91%
  209490 requests in 1.00m, 48.37MB read
Requests/sec:   3485.86
Transfer/sec:    824.24KB
```

#### 800 connections

```
Running 1m test @ http://localhost:9292
  12 threads and 800 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   111.83ms   63.05ms 492.30ms   61.15%
    Req/Sec   296.34    166.81   730.00     62.88%
  212895 requests in 1.00m, 49.18MB read
Requests/sec:   3542.40
Transfer/sec:    837.90KB
```